changing a location of images
Now, folked hugo theme can find the images in the corresponding posts directory.
All markdown files in the posts directory was moved to new directorys in the posts dir and most images were moved to the corresponding directory.
All images are now resized.
Some images are updated for new style.